Fix issue where adjacent comments were skipped during canonicalization
Addresses issue #50

If comments were adjacent in a document with no whitespace between them, then for each pair of comments the second comment would be skipped due to the underlying slice being modified as it was being ranged over.

For example, the following document:

```xml
<a><!-- comment1 --><!-- comment2 --><!-- comment 3 --></a>
```

Would be incorrectly canonicalized to:

```xml
<a><!-- comment2 --></a>
```

This patch fixes that issue and adds a new test case to cover it.